chore: move manager test files
