+++ lib/manager/helm-values/__fixtures__/default_chart_init_values.yaml
