+++ lib/manager/helm-values/__fixtures__/multi_and_nested_image_values.yaml
