+++ lib/manager/helm-values/extract.spec.ts
@@ -1,14 +1,13 @@
 import { readFileSync } from 'fs';
-
-import { extractPackageFile } from '../../../lib/manager/helm-values/extract';
+import { extractPackageFile } from './extract';
 
 const helmDefaultChartInitValues = readFileSync(
-  'test/manager/helm-values/_fixtures/default_chart_init_values.yaml',
+  'lib/manager/helm-values/__fixtures__/default_chart_init_values.yaml',
   'utf8'
 );
 
 const helmMultiAndNestedImageValues = readFileSync(
-  'test/manager/helm-values/_fixtures/multi_and_nested_image_values.yaml',
+  'lib/manager/helm-values/__fixtures__/multi_and_nested_image_values.yaml',
   'utf8'
 );
